Revert dual stack changes for later PR
