+++ entrypoint/ipv6only/ipv6only_test.go
@@ -3,22 +3,41 @@ package ipv6only
 import (
 	"fmt"
 
+	"github.com/rancher/distros-test-framework/pkg/assert"
 	"github.com/rancher/distros-test-framework/pkg/testcase"
+	"github.com/rancher/distros-test-framework/pkg/testcase/support"
+	"github.com/rancher/distros-test-framework/shared"
 
 	. "github.com/onsi/ginkgo/v2"
 )
 
-var _ = Describe("Test ipv6 only cluster:", func() {
-	It("Start Up with no issues", func() {
-		testcase.TestBuildIPv6OnlyCluster(cluster)
+var _ = Describe("Test ipv6 only cluster:", Ordered, func() {
+	BeforeAll(func() {
+		support.BuildIPv6OnlyCluster(cluster)
 	})
 
 	It("Install product on ipv6 only nodes", func() {
 		testcase.TestIPv6Only(cluster, awsClient)
 	})
 
-	// TODO: validates
+	It("Validates Nodes", func() {
+		testcase.TestAirgapClusterNodeStatus(
+			cluster,
+			assert.NodeAssertReadyStatus(),
+			nil,
+		)
+	})
 
+	It("Validates Pods", func() {
+		testcase.TestAirgapClusterPodStatus(
+			cluster,
+			assert.PodAssertRestart(),
+			assert.PodAssertReady())
+	})
+
+	AfterAll(func()  {
+		shared.DisplayAirgapClusterDetails(cluster)
+	})
 })
 
 var _ = AfterEach(func() {
