--- modules/dualstack/instance/providers.tf
@@ -1,3 +1,0 @@
-provider "aws" {
-    region = "${var.region}"
-}
