--- modules/dualstack/main.tf
@@ -1,22 +1,0 @@
-module "instance" {
-   source               = "./instance"
-   aws_ami              = var.aws_ami
-   aws_user             = var.aws_user
-   ec2_instance_class   = var.ec2_instance_class
-   region               = var.region
-   vpc_id               = var.vpc_id
-   bastion_subnets      = var.bastion_subnets
-   availability_zone    = var.availability_zone
-   sg_id                = var.sg_id
-   subnets              = var.subnets
-   enable_public_ip     = var.enable_public_ip
-   enable_ipv6          = var.enable_ipv6
-   key_name             = var.key_name
-   access_key           = var.access_key
-   resource_name        = var.resource_name
-   volume_size          = var.volume_size
-   no_of_bastion_nodes  = var.no_of_bastion_nodes
-   no_of_server_nodes   = var.no_of_server_nodes
-   no_of_worker_nodes   = var.no_of_worker_nodes
-   product              = var.product 
-}
