+++ modules/rke2/worker/variables.tf
@@ -17,10 +17,10 @@ variable "password" {
 }
 variable "region" {}
 variable "resource_name" {}
-variable "install_version" {}
+variable "rke2_version" {}
 variable "install_mode" {}
 variable "install_method" {}
-variable "install_channel" {}
+variable "rke2_channel" {}
 variable "sg_id" {}
 variable "subnets" {}
 variable "username" {
@@ -28,6 +28,12 @@ variable "username" {
 }
 variable "vpc_id" {}
 variable "worker_flags" {}
+variable "enable_public_ip" {
+  default = true
+}
+variable "enable_ipv6" {
+  default = false
+}
 variable "create_eip" {
   default = false
 }
